refactor: Simplify projects structure to only 3 projects
