+++ src/CSharpLeetCode.Benchmarks.Console/CSharpLeetCode.Benchmarks.Console.csproj
@@ -1,11 +1,11 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <ItemGroup>
-    <ProjectReference Include="..\Problem0001.TwoSum.Library\Problem0001.TwoSum.Library.csproj" />
+    <PackageReference Include="BenchmarkDotNet" Version="0.14.0" />
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="BenchmarkDotNet" Version="0.14.0" />
+    <ProjectReference Include="..\CSharpLeetCode.Solutions.Library\CSharpLeetCode.Solutions.Library.csproj" />
   </ItemGroup>
 
   <PropertyGroup>
