+++ src/CSharpLeetCode.Benchmarks.Console/Program.cs
@@ -1,5 +1,5 @@
 ﻿using BenchmarkDotNet.Running;
-using Benchmarks.Console.Benchmarks;
+using CSharpLeetCode.Benchmarks.Console.Benchmarks;
 
 // Select the benchmark to run within the <>.
 var summary = BenchmarkRunner.Run<Problem0001TwoSum>();
