+++ src/CSharpLeetCode.Solutions.Library/Interfaces/IProblem0001TwoSum.cs
@@ -1,0 +1,6 @@
+namespace CSharpLeetCode.Solutions.Library.Interfaces;
+
+public interface IProblem0001TwoSum
+{
+    int[] TwoSum(int[] nums, int target);
+}
