--- src/Problem0001.TwoSum.Library/Interfaces/ISolution.cs
@@ -1,6 +1,0 @@
-namespace Problem0001.TwoSum.Library.Interfaces;
-
-public interface ISolution
-{
-    int[] TwoSum(int[] nums, int target);
-}
